--- a/src/error.rs
+++ b/src/error.rs
@@ -47,10 +47,8 @@ pub enum Error {
     /// Operation not supported or unimplemented on this platform.
     NotSupported,
 
-    /* Consider adding this on next breaking change (v0.6?)
-    /// The device returned a malformed descriptor
+    /// The device returned a malformed descriptor.
     BadDescriptor,
-    */
 
     /// Other error.
     Other,
@@ -73,7 +71,7 @@ impl Error {
             Error::Interrupted => "System call interrupted (perhaps due to signal)",
             Error::NoMem => "Insufficient memory",
             Error::NotSupported => "Operation not supported or unimplemented on this platform",
-            //Error::BadDescriptor => "Malformed descriptor",
+            Error::BadDescriptor => "Malformed descriptor",
             Error::Other => "Other error",
         }
     }