As 0.6 release seems like obvious, I took a look at the code and did minor rework so those changes could go in, if approved, on the same release.

Take a look at the proposed changes.