@a1ien 2020-06-12
I would not want to add additional dependencies. Can we go without derive_more