@otavio 2020-06-12
> I would not want to add additional dependencies. Can we go without derive_more

We can, for sure. If desired, I can drop it from the branch.

As a reason to keep it is that `derive_more` only depends on `syn` and it is a very common dependency. Also, as I did enable just the macros we currently use.

I personally prefer to reduce the amount of code to maintain and to prefer commonly used dependencies, so we 'reuse' the previously built dependencies. In this scenario, the built time would be the same.

So please, what solution do you prefer?