@pheki 2020-06-13
Just tested it, and on my machine (which is kinda slow) it adds 10s to the first debug build and about 50s in first release build when adding derive_more directly for my crate (its not on my dependency tree).

Without `derive_more`:
```
$ cargo clean
$ cargo build
Finished dev [unoptimized + debuginfo] target(s) in 58.27s
$ cargo clean --release
$ cargo build --release
Finished release [optimized] target(s) in 2m 36s
```
With `derive_more = "0.99"`:
```
$ cargo clean
$ cargo build
Finished dev [unoptimized + debuginfo] target(s) in 1m 07s
$ cargo clean --release
$ cargo build --release
Finished release [optimized] target(s) in 3m 29s
```

For reference:
```
[dependencies]
rusb = "0.5"
structopt = "0.3"
crossbeam-channel = "0.4"
lazy_static = "1.4"
# derive_more = "0.99"
```

### Just my opinion

While I agree that adding dependencies is not necessarily bad, specially "shared" ones, I think that the maintenance burden (e.g. updating the library when 0.100 comes out) isn't worth the trouble in this case.