@otavio 2020-06-14
Already. I removed derive_more use. Please check it again @pheki @a1ien.